Agregue la clase OFad y sus respectivos VO

+++ Zara/ejbModule/server/VO/articulos/ArtHogarVO.java
@@ -11,7 +11,7 @@ public class ArtHogarVO extends ArticuloVO {
 		
 	}
 	
-	public ArtHogarVO(int ref, String desc, String linea, double pL, double pO, 
+	public ArtHogarVO(long ref, String desc, String linea, double pL, double pO, 
 			 String color, String sec, int stock, String nom, 
 			 String comp, String med, String cat) {
